Change name regexes to accomodate names with spaces and periods.
Hi All, 

I tried to register for the ruby mailing list and noticed that there was a regex that filters for name validity. I've updated the regular expression so it allows for person's who have spaces in their names or periods. Now I (Earl St Sauver) and any number of other people (The Mc. Anthing's) can register. 

~Earl