Use create_table instead of create_table? for schema
When a table already exists, create_table? silently does nothing.
So after changing schema.rb and running 'rake db:setup' the database
might still have the outdated schema and not the expected one.
Using create_table instead raises an exception, providing immediate
feedback that the database might not be in the expected state.